+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased 5.1.7]
 ## Features
 ## Improvements
+- SALE ORDER : line not editable if availability request = true
+
 ## Bug Fixes